feat(refactor): move projectsData to data folder
